Add tests
* Move things around so that highlightHtml is in its own folder
