+++ src/components/LogRow/BaseRow/Row.test.tsx
@@ -81,31 +81,6 @@ describe("row", () => {
   });
 
   describe("search", () => {
-    it("a search term highlights the matching text", () => {
-      const regexp = /Test/i;
-      renderWithRouterMatch(
-        <Row {...rowProps} searchTerm={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.getByDataCy("highlight")).toHaveTextContent("Test");
-    });
-    it("should preserve case sensitivity when applying a search", () => {
-      let regexp = /test/i;
-      const { rerender } = renderWithRouterMatch(
-        <Row {...rowProps} searchTerm={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.getByDataCy("highlight")).toHaveTextContent("Test");
-      regexp = /test/;
-      rerender(
-        <Row {...rowProps} searchTerm={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.queryByDataCy("highlight")).toBeNull();
-    });
     it("should highlight matching text if it is within range", () => {
       const regexp = /Test/i;
       renderWithRouterMatch(
@@ -137,83 +112,6 @@ describe("row", () => {
         </Row>
       );
       expect(screen.queryByDataCy("highlight")).not.toBeInTheDocument();
-    });
-    it("should not highlight the content in hrefs", () => {
-      const logLine = `highlight me <a href="https://donthighlightme.com">highlight me</a> highlight me`;
-      const regexp = /highlight/gi;
-      renderWithRouterMatch(
-        <Row {...rowProps} searchTerm={regexp}>
-          {logLine}
-        </Row>
-      );
-      expect(screen.queryAllByDataCy("highlight")).toHaveLength(3);
-      expect(
-        screen.getByRole("link", { name: "highlight me" })
-      ).toHaveAttribute("href", "https://donthighlightme.com");
-    });
-  });
-
-  describe("highlights", () => {
-    it("highlighted terms should highlight the matching text", () => {
-      const regexp = /Test/i;
-      renderWithRouterMatch(
-        <Row {...rowProps} highlightRegex={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.getByDataCy("highlight")).toHaveTextContent("Test");
-    });
-    it("should highlight every matching term on a line", () => {
-      const regexp = /Test|Log/gi;
-      renderWithRouterMatch(
-        <Row {...rowProps} highlightRegex={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.queryAllByDataCy("highlight")).toHaveLength(2);
-      screen.getAllByDataCy("highlight").forEach((highlight) => {
-        expect(highlight).toHaveTextContent(/Test|Log/i);
-      });
-    });
-    it("should deduplicate highlights and searches on the same string", () => {
-      const regexp = /Test/i;
-      renderWithRouterMatch(
-        <Row {...rowProps} highlightRegex={regexp} searchTerm={regexp}>
-          {testLog}
-        </Row>
-      );
-      expect(screen.queryAllByDataCy("highlight")).toHaveLength(1);
-      expect(screen.getByDataCy("highlight")).toHaveTextContent("Test");
-    });
-    it("should show both highlights and searches if they are on the same line", () => {
-      const searchRegex = /Test/i;
-      const highlightRegex = /(Log)/i;
-      renderWithRouterMatch(
-        <Row
-          {...rowProps}
-          highlightRegex={highlightRegex}
-          searchTerm={searchRegex}
-        >
-          {testLog}
-        </Row>
-      );
-      expect(screen.queryAllByDataCy("highlight")).toHaveLength(2);
-      screen.getAllByDataCy("highlight").forEach((highlight) => {
-        expect(highlight).toHaveTextContent(/Test|Log/i);
-      });
-    });
-    it("should not highlight the content in hrefs", () => {
-      const logLine = `highlight me <a href="https://donthighlightme.com">highlight me</a> highlight me`;
-      const highlightRegex = /(highlight)/gi;
-      renderWithRouterMatch(
-        <Row {...rowProps} highlightRegex={highlightRegex}>
-          {logLine}
-        </Row>
-      );
-      expect(screen.queryAllByDataCy("highlight")).toHaveLength(3);
-      expect(
-        screen.getByRole("link", { name: "highlight me" })
-      ).toHaveAttribute("href", "https://donthighlightme.com");
     });
   });
 });
